+++ R/ppc-intervals.R
@@ -7,13 +7,12 @@
 #' @family PPCs
 #'
 #' @template args-y-yrep
+#' @template args-prob-prob_outer
 #' @param x A numeric vector the same length as \code{y} to use as the x-axis
 #'   variable. For example, \code{x} could be a predictor variable from a
 #'   regression model, a time variable for time-series models, etc. If \code{x}
 #'   is missing or NULL, then \code{1:length(y)} is used for the x-axis.
 #' @param ... Currently unused.
-#' @param prob A value between 0 and 1 indicating the desired probability mass
-#'   to include in the \code{yrep} intervals. The default is 0.9.
 #' @param alpha,size,fatten Arguments passed to geoms. For ribbon plots
 #'   \code{alpha} and \code{size} are passed to
 #'   \code{\link[ggplot2]{geom_ribbon}}. For interval plots \code{size} and
@@ -109,8 +108,8 @@ NULL
 
 #' @rdname PPC-intervals
 #' @export
-ppc_intervals <- function(y, yrep, x = NULL, ..., prob = 0.9, size = 1,
-                          fatten = 3) {
+ppc_intervals <- function(y, yrep, x = NULL, ..., prob = 0.5, prob_outer = 0.9,
+                          size = 1, fatten = 3) {
   check_ignored_arguments(...)
 
   data <- ppc_intervals_data(
@@ -118,7 +117,8 @@ ppc_intervals <- function(y, yrep, x = NULL, ..., prob = 0.9, size = 1,
       yrep = yrep,
       x = x,
       group = NULL,
-      prob = prob
+      prob = prob,
+      prob_outer = prob_outer
   )
 
   .ppc_intervals(
@@ -143,7 +143,8 @@ ppc_intervals_grouped <- function(y,
                                   group,
                                   facet_args = list(),
                                   ...,
-                                  prob = 0.9,
+                                  prob = 0.5,
+                                  prob_outer = 0.9,
                                   size = 1,
                                   fatten = 3) {
   check_ignored_arguments(...)
@@ -157,7 +158,8 @@ ppc_intervals_grouped <- function(y,
     yrep = yrep,
     x = x,
     group = group,
-    prob = prob
+    prob = prob,
+    prob_outer = prob_outer
   )
 
   .ppc_intervals(
@@ -178,7 +180,8 @@ ppc_ribbon <- function(y,
                        yrep,
                        x = NULL,
                        ...,
-                       prob = 0.9,
+                       prob = 0.5,
+                       prob_outer = 0.9,
                        alpha = 0.33,
                        size = 0.25) {
   check_ignored_arguments(...)
@@ -210,7 +213,8 @@ ppc_ribbon_grouped <- function(y,
                                group,
                                facet_args = list(),
                                ...,
-                               prob = 0.9,
+                               prob = 0.5,
+                               prob_outer = 0.9,
                                alpha = 0.33,
                                size = 0.25) {
   check_ignored_arguments(...)
@@ -224,7 +228,8 @@ ppc_ribbon_grouped <- function(y,
     yrep = yrep,
     x = x,
     group = group,
-    prob = prob
+    prob = prob,
+    prob_outer = prob_outer
   )
 
   .ppc_intervals(
@@ -241,9 +246,11 @@ ppc_ribbon_grouped <- function(y,
 
 #' @rdname PPC-intervals
 #' @export
-ppc_intervals_data <- function(y, yrep, x = NULL, group = NULL, prob = 0.9, ...) {
+ppc_intervals_data <- function(y, yrep, x = NULL, group = NULL,
+                               prob = 0.5, prob_outer = 0.9, ...) {
   check_ignored_arguments(...)
-  .ppc_intervals_data(y = y, yrep = yrep, x = x, group = group, prob = prob)
+  .ppc_intervals_data(y = y, yrep = yrep, x = x, group = group,
+                      prob = prob, prob_outer = prob_outer)
 }
 
 
@@ -259,9 +266,14 @@ label_x <- function(x) {
   if (missing(x)) "Index" else NULL
 }
 
-.ppc_intervals_data <- function(y, yrep, x = NULL, group = NULL, prob = 0.8) {
+.ppc_intervals_data <- function(y, yrep, x = NULL, group = NULL,
+                                prob = 0.5, prob_outer = 0.9) {
   grouped <- !is.null(group)
   stopifnot(prob > 0 && prob < 1)
+  stopifnot(prob_outer > 0 && prob_outer <= 1)
+  probs <- sort(c(prob, prob_outer))
+  prob <- probs[1]
+  prob_outer <- probs[2]
 
   y <- validate_y(y)
   yrep <- validate_yrep(yrep, y)
@@ -283,16 +295,19 @@ label_x <- function(x) {
   }
 
   grouped_d <- dplyr::group_by(molten_reps, !!! group_vars)
-  alpha <- (1 - prob) / 2
+  alpha <- (1 - probs) / 2
   probs <- sort(c(alpha, 0.5, 1 - alpha))
 
   val_col <- sym("value")
   dplyr::ungroup(dplyr::summarise(
     grouped_d,
-    prob = prob,
-    lo = quantile(!! val_col, prob = probs[1]),
-    mid = quantile(!! val_col, prob = probs[2]),
-    hi = quantile(!! val_col, prob = probs[3])
+    outer_width = prob_outer,
+    inner_width = prob,
+    ll = quantile(!! val_col, prob = probs[1]),
+    l  = quantile(!! val_col, prob = probs[2]),
+    m  = quantile(!! val_col, prob = probs[3]),
+    h  = quantile(!! val_col, prob = probs[4]),
+    hh = quantile(!! val_col, prob = probs[5])
   ))
 }
 
@@ -317,14 +332,19 @@ label_x <- function(x) {
     data = data,
     mapping = aes_(
       x = ~ x,
-      y = ~ mid,
-      ymin = ~ lo,
-      ymax = ~ hi
+      y = ~ m,
+      ymin = ~ l,
+      ymax = ~ h
     )
   )
 
   if (style == "ribbon") {
     graph <- graph +
+      geom_ribbon(
+        aes_(color = "yrep", fill = "yrep", ymin = ~ ll, ymax = ~ hh),
+        alpha = alpha,
+        size = size
+      ) +
       geom_ribbon(
         aes_(color = "yrep", fill = "yrep"),
         alpha = alpha,
@@ -341,6 +361,13 @@ label_x <- function(x) {
       )
   } else {
     graph <- graph +
+      geom_pointrange(
+        mapping = aes_(color = "yrep", fill = "yrep", ymin = ~ ll, ymax = ~ hh),
+        shape = 21,
+        alpha = alpha,
+        size = size,
+        fatten = fatten
+      ) +
       geom_pointrange(
         mapping = aes_(color = "yrep", fill = "yrep"),
         shape = 21,